fix(index): enforce bounded Property repair paging contract
