+++ pkg/index/inverted/native_nidx01b_boundary_test.go
@@ -75,6 +75,8 @@ var nativeReaderSurface = []string{
 	"Reader.VisibleDocCount",
 	"Reader.VisitLiveDocuments",
 	"Reader.VisitSelectedDocuments",
+	"RepairCursor",
+	"RepairPageRequest",
 	"RepairTupleRow",
 	"StoredDocument",
 }
